refactor merge obj
